don't want warrior to get out of the logic side (for now)

+++ include/player.hpp
@@ -10,6 +10,6 @@ class Player : public Warrior {
 private:
   
 public:
-  Player(double xcor,double ycor)
+  Player(float xcor,float ycor)
     : Warrior(xcor,ycor){};
 };
