+++ include/warrior.hpp
@@ -17,7 +17,7 @@ protected:
    * angle: angle of the warrior's velocity in radians
    * speed: magnitude of the warrior's velocity in pixels per microsecond
    */
-  double xcor,ycor,angle,speed;
+  float xcor,ycor,angle,speed;
 
 
   /*
@@ -26,7 +26,7 @@ protected:
     @param xcor initial x-coordinate of the warrior.
     @param ycor initial y-coordinate of the warrior.
    */
-  Warrior(double xcor,double ycor)
+  Warrior(float xcor,float ycor)
   {this->xcor=xcor;this->ycor=ycor;this->angle=M_PI/4;this->speed=0.0001;}
 public:
   
@@ -35,7 +35,7 @@ public:
 
     @return the warrior's x-coordinate
   */
-  double get_xcor(){return xcor;}
+  float get_xcor(){return xcor;}
 
   
   /*
@@ -43,7 +43,7 @@ public:
 
     @return the warrior's y-coordinate
   */
-  double get_ycor(){return ycor;}
+  float get_ycor(){return ycor;}
 
   /*
      Sets the coordinates of the warrior to what they would be after the
